Lifted SelectBox's states up
